+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Ability to specify how to match values together ([#15](https://github.com/khalyomede/reorder-before-after/issues/15)).
 
+### Breaking
+
+- The "BadOutOfCallbackException" has been renamed to "InvalidOutOfCallbackException" to match the other similar exceptions prefixes ([#17](https://github.com/khalyomede/reorder-before-after/issues/17)).
+
 ## [0.2.0] - 2023-01-14
 
 ### Added